+++ com_payzen/classes/PayzenField.php
@@ -1,32 +1,19 @@
 <?php
 /**
- * PayZen V2-Payment Module version 2.1.0 for VirtueMart 3.x. Support contact : support@payzen.eu.
+ * Copyright © Lyra Network.
+ * This file is part of PayZen plugin for VirtueMart. See COPYING.md for license details.
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- *
- * @author    Lyra Network (http://www.lyra-network.com/)
- * @copyright 2014-2017 Lyra Network and contributors
- * @license   http://www.gnu.org/licenses/old-licenses/gpl-2.0.html  GNU General Public License (GPL v2)
- * @category  payment
- * @package   payzen
- */
+ * @author    Lyra Network (https://www.lyra.com/)
+ * @copyright Lyra Network
+ * @license   http://www.gnu.org/licenses/old-licenses/gpl-2.0.html GNU General Public License (GPL v2)
+*/
+
+defined('_JEXEC') or die('Restricted access');
 
 if (! class_exists('PayzenField', false)) {
 
     /**
-     * Class representing a form field to send to the payment platform.
+     * Class representing a form field to send to the payment gateway.
      */
     class PayzenField
     {
@@ -121,7 +108,7 @@ if (! class_exists('PayzenField', false)) {
         public function setValue($value)
         {
             $value = ($value === null) ? null : (string) $value;
-            // we save value even if invalid but we return "false" as warning
+            // We save value even if invalid but we return "false" as warning.
             $this->value = $value;
 
             return $this->isValid();